Extracted google secret shared context to it's own file
