+++ spec/secrets/google_cloud/env_loader_spec.rb
@@ -1,29 +1,10 @@
 
 require 'on_container/secrets/google_cloud/env_loader'
+require_relative '../shared/contexts/with_google_secret_manager'
 
 RSpec.describe OnContainer::Secrets::GoogleCloud::EnvLoader, type: :env_spec do
   shared_context 'ENV with vars ending with "_GOOGLE_CLOUD_SECRET"' do
     let(:example_env_vars) { { FOO_GOOGLE_CLOUD_SECRET: 'foo' } }
-  end
-
-  shared_context 'Google::Cloud::SecretManager is loaded' do
-    around do |example|
-      module ::Google
-        module Cloud
-          module SecretManager
-            def self.secret_manager_service
-              'bar'
-            end
-          end
-        end
-      end
-  
-      example.run
-      
-      Google::Cloud.send(:remove_const, :SecretManager)
-      Google.send(:remove_const, :Cloud)
-      Object.send(:remove_const, :Google)
-    end
   end
 
   describe '#env_keys' do
@@ -65,8 +46,8 @@ RSpec.describe OnContainer::Secrets::GoogleCloud::EnvLoader, type: :env_spec do
       end
     end
 
-    context 'when Google::Cloud::SecretManager is loaded' do
-      include_context 'Google::Cloud::SecretManager is loaded'
+    context 'with Google::Cloud::SecretManager loaded' do
+      include_context 'with Google::Cloud::SecretManager loaded'
 
       it 'returns true' do
         expect(subject.env_keys?).to be false
@@ -85,8 +66,8 @@ RSpec.describe OnContainer::Secrets::GoogleCloud::EnvLoader, type: :env_spec do
         end
       end
 
-      context 'when Google::Cloud::SecretManager is loaded' do
-        include_context 'Google::Cloud::SecretManager is loaded'
+      context 'with Google::Cloud::SecretManager loaded' do
+        include_context 'with Google::Cloud::SecretManager loaded'
 
         it 'does not call the fetcher' do
           expect(fetcher_class).not_to receive(:perform!)
@@ -105,8 +86,9 @@ RSpec.describe OnContainer::Secrets::GoogleCloud::EnvLoader, type: :env_spec do
         end
       end
 
-      context 'when Google::Cloud::SecretManager is loaded' do
-        include_context 'Google::Cloud::SecretManager is loaded'
+      context 'with Google::Cloud::SecretManager loaded' do
+        include_context 'with Google::Cloud::SecretManager loaded'
+
         let(:test_secret_data) { { 'UNO' => 'ONE' } }
 
         it 'merges the fetched data to ENV' do
